Adjust styles and fix hover and focus issues
