+++ app/layout.tsx
@@ -2,14 +2,23 @@ import type { Metadata } from 'next';
 import './globals.css';
 import Link from 'next/link';
 import Image from 'next/image';
-import { M_PLUS_1p } from 'next/font/google';
+import { M_PLUS_1p, M_PLUS_1, Source_Sans_3 } from 'next/font/google';
 
 import { logo_dark } from './res/logo';
+import { clsx } from './utils';
 
-const mplus = M_PLUS_1p({
+const mplusp = M_PLUS_1p({
   variable: '--font-mplus-1p',
   subsets: ['latin'],
   weight: ['100', '300', '400', '500', '700', '800', '900'],
+});
+const mplus = M_PLUS_1({
+  variable: '--font-mplus-1',
+  subsets: ['latin'],
+});
+const source = Source_Sans_3({
+  variable: '--font-source-sans',
+  subsets: ['latin'],
 });
 
 export const metadata: Metadata = {
@@ -27,18 +36,22 @@ export default function RootLayout({
   children: React.ReactNode;
 }>) {
   return (
-    <html lang="ja" className={`scroll-container${(development || preview) ? ' has-[#theme-dark:checked]:[color-scheme:dark] has-[#theme-dark:checked]:[--tw-color-scheme:dark] has-[#theme-light:checked]:[color-scheme:light] has-[#theme-light:checked]:[--tw-color-scheme:light]' : ''}`}>
-      <body className={`min-h-screen ${mplus.variable} font-sans font-light antialiased bg-[url('https://picsum.photos/seed/eternalcore/1920/1200?grayscale')] bg-cover`}>
+    <html lang="ja" className={clsx(`scroll-container`, (development || preview) && 'has-[#theme-dark:checked]:[color-scheme:dark] has-[#theme-dark:checked]:[--tw-color-scheme:dark] has-[#theme-light:checked]:[color-scheme:light] has-[#theme-light:checked]:[--tw-color-scheme:light]')}>
+      <body className={`min-h-screen bg-[url('https://picsum.photos/seed/eternalcore/1920/1200?grayscale')] bg-cover font-sans font-light antialiased ${mplus.variable} ${mplusp.variable} ${source.variable}`}>
 
         <header className="flex flex-row fixed inset-x-0 top-0 z-10 text-white backdrop-blur-sm backdrop-brightness-50 border-b-white border-b-2 max-md:flex-col max-md:items-center max-md:not-at-page-top:-translate-y-24 transition-all delay-200 duration-500">
           <Link href="/" className="h-24 md:not-at-page-top:h-16 transition-all delay-200 duration-500">
             <Image src={logo_dark} alt="二部ソフトウェア研究部" className="px-16 py-4 h-full w-fit object-contain md:not-at-page-top:py-2 md:not-at-page-top:px-24 transition-all delay-200 duration-500" />
           </Link>
-          <ul className="nav h-24 ms-auto flex flex-row gap-1 max-md:h-12 *:max-md:flex-1 max-md:w-full border-white max-md:border-t-[1px] not-at-page-top:h-16 transition-all  delay-200 duration-500">
-            <li><Link href="/about">About Us</Link></li>
-            <li><Link href="/news">News</Link></li>
-            <li><Link href="/blog">Blog</Link></li>
-            <li><Link href="/contact">Contact</Link></li>
+          <ul className="h-24 ms-auto flex flex-row gap-1 max-md:h-12 *:max-md:flex-1 max-md:w-full border-white max-md:border-t-[1px] not-at-page-top:h-16 transition-all delay-200 duration-500">
+            {([
+              ['/about', 'About Us'],
+              ['/news', 'News'],
+              ['/blog', 'Blog'],
+              ['/contact', 'Contact'],
+            ] as const).map(([href, name]) => (<li key={href}>
+              <Link href={href} className="block min-w-max h-full px-8 text-center content-center max-md:px-0 max-lg:px-4 text-white hocus:text-black transition-all duration-500 ease-in-out [box-shadow:inset_0_var(--sw)_white] [--sw:-2px] hocus:[--sw:calc(-1*theme('spacing.24'))] max-md:hocus:[--sw:calc(-1*theme('spacing.12'))] device-touch:[--sw:-4px] device-touch:active:[--sw:-1px] device-touch:duration-200">{name}</Link>
+            </li>))}
           </ul>
         </header>
 
@@ -84,46 +97,44 @@ export default function RootLayout({
           </a>
         </footer>
 
-        {preview && <label className="fixed bottom-4 right-4 w-min self-center p-2 bg-[repeating-linear-gradient(135deg,black_0px,black_20px,yellow_20px,yellow_40px)] transition-all duration-500 has-[#preview-tip:checked]:translate-x-full has-[#preview-tip:checked]:opacity-20 cursor-pointer">
-          <input type="checkbox" id="preview-tip" name="hide-preview-tip" className="sr-only" />
-          <span className="flex flex-row items-center self-center bg-slate-600 px-2 py-1 text-white">
-            <span className="icon-material-symbols-light-construction text-4xl me-2"></span>
-            <span className="first-line:text-lg first-line:leading-none text-xs w-min">Preview<br />Under Construction</span>
-          </span>
+        {(preview) && <label className="fixed bottom-4 right-4 w-max self-center p-2 bg-[repeating-linear-gradient(135deg,black_0px,black_20px,yellow_20px,yellow_40px)] transition-all duration-500 has-[#preview-tip:checked]:translate-x-full has-[#preview-tip:checked]:opacity-20 cursor-pointer">
+          <input type="checkbox" id="preview-tip" name="hide-preview-tip" className="absolute inset-0 appearance-none cursor-pointer" />
+          <div className="flex flex-row items-center self-center bg-slate-600 px-2 py-1 text-white">
+            <span className="icon-mslight-construction text-4xl me-2"></span>
+            <div>
+              <div className="text-lg leading-none">Preview</div>
+              <div className="text-xs">Under Construction</div>
+            </div>
+          </div>
         </label>}
 
-        {false && <div className="fixed inset-0 m-auto size-max text-4xl select-none">
+        {(false) && <div className="fixed inset-0 m-auto size-max text-4xl select-none">
           <span className="sm:after:content-['sm'] md:after:content-['md'] lg:after:content-['lg'] xl:after:content-['xl']" />
           <span> / </span>
           <span className="max-sm:after:content-['max-sm'] max-md:after:content-['max-md'] max-lg:after:content-['max-lg'] max-xl:after:content-['max-xl']" />
         </div>}
 
-        {(development || preview) && <div className="fixed inset-y-0 right-4 m-auto size-10 p-1 bg-black text-white bg-opacity-75 [&_input]:sr-only group/theme">
-          <label className="sr-only group-has-[#theme-light:checked]/theme:not-sr-only cursor-pointer">
-            <input type="radio" name="color-scheme" id="theme-system" defaultChecked />
-            <span className="size-8 icon-mslight-brightness-auto-outline"></span>
-          </label>
-          <label className="sr-only group-has-[#theme-system:checked]/theme:not-sr-only cursor-pointer">
-            <input type="radio" name="color-scheme" id="theme-dark" />
-            <span className="size-8 icon-mslight-light-mode-outline"></span>
-          </label>
-          <label className="sr-only group-has-[#theme-dark:checked]/theme:not-sr-only cursor-pointer">
-            <input type="radio" name="color-scheme" id="theme-light" />
-            <span className="size-8 icon-mslight-dark-mode-outline"></span>
-          </label>
+        {(development || preview) && <div className="flex flex-col fixed inset-y-0 right-4 m-auto size-min p-1 bg-black/80 text-white rounded-full -space-y-1 before:absolute before:inset-x-0 before:size-8 before:m-1 before:rounded-full before:bg-white/20 has-[#theme-light:checked]:before:top-0 has-[#theme-system:checked]:before:top-[1.75rem] has-[#theme-dark:checked]:before:top-[3.5rem] before:transition-all">
+          {([
+            ['theme-light', 'icon-mslight-light-mode-outline', 'Light Mode'],
+            ['theme-system', 'icon-mslight-display-settings-outline', 'No Override', true],
+            ['theme-dark', 'icon-mslight-dark-mode-outline', 'Dark Mode'],
+          ] as const).map(([name, icon, desc, checked]) => (
+            <label key={name} className="size-8 p-1.5 rounded-full cursor-pointer relative group">
+              <input type="radio" name="color-scheme" className="absolute inset-0 rounded-full appearance-none" id={name} defaultChecked={checked} />
+              <span className="absolute inset-y-0 -left-2 m-auto px-2 size-max -translate-x-full invisible group-hover:visible bg-black/80 rounded-full text-sm device-touch:group-active:delay-0 device-touch:group-has-[:checked]:delay-1000 select-none">{desc}</span>
+              <span className={`size-full ${icon} group-hover:scale-125 transition-all`} />
+            </label>
+          ))}
         </div>}
 
-        {(development || preview) && <label className="fixed inset-y-0 left-4 m-auto size-max min-w-40 p-1 bg-black text-white bg-opacity-75 transition-all duration-500 has-[#scroll-tip:checked]:-translate-x-full has-[#scroll-tip:checked]:opacity-20 has-[#scroll-tip:checked]:pe-4 cursor-pointer">
-          <input type="checkbox" id="scroll-tip" name="hide-scroll-tip" className="sr-only" defaultChecked />
+        {(development || preview) && <label className="flex flex-col gap-1 fixed inset-y-0 left-4 m-auto size-max min-w-40 p-1 bg-black/80 text-white transition-all duration-500 has-[#scroll-tip:checked]:-translate-x-full has-[#scroll-tip:checked]:opacity-20 has-[#scroll-tip:checked]:pe-4 cursor-pointer">
+          <input type="checkbox" id="scroll-tip" name="hide-scroll-tip" className="absolute inset-0 appearance-none cursor-pointer" defaultChecked />
 
-          <div className="w-full relative after:absolute after:inset-x-0 after:bottom-0 after:h-[1px] after:bg-white after:origin-left after:scale-x-[--scroll-pos]">scroll-position</div>
-          <div className="w-full relative after:absolute after:right-0 after:left-1/2 after:bottom-0 after:h-[1px] after:bg-white after:origin-left after:scale-x-[--scroll-vel]">scroll-velocity</div>
-          <div className="not-scrolling:bg-white not-scrolling:text-black">state: stopped</div>
-          <div className="scrolling:bg-white scrolling:text-black">state: scrolling</div>
-          <div className="scrolling-up:bg-white scrolling-up:text-black">direction: up</div>
-          <div className="scrolling-down:bg-white scrolling-down:text-black">direction:-down</div>
-          <div className="at-page-top:bg-white at-page-top:text-black">scroll pos: top</div>
-          <div className="at-page-bottom:bg-white at-page-bottom:text-black">scroll pos: bottom</div>
+          <div className="w-full relative after:absolute after:inset-x-0 after:bottom-0 after:h-px after:bg-current after:origin-left after:scale-x-[--scroll-pos]">scroll-position</div>
+          <div className="w-full relative after:absolute after:right-0 after:left-1/2 after:bottom-0 after:h-px after:bg-current after:origin-left after:scale-x-[--scroll-vel]">scroll-velocity</div>
+          <div className="scrolling-up:after:content-['up'] scrolling-down:after:content-['down'] not-scrolling:after:content-['stopped'] scrolling:text-black scrolling:bg-white">scroll dir: </div>
+          <div className="after:content-['middle'] at-page-top:after:content-['top'] at-page-top:text-black at-page-top:bg-white at-page-bottom:after:content-['bottom'] at-page-bottom:text-black at-page-bottom:bg-white">scroll pos: </div>
         </label>}
       </body>
     </html>
